app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
qdrant-client==1.7.0
openai==1.3.5
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.0
google-cloud-translate==3.11.3
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
tiktoken==0.5.2
langchain==0.0.339
langchain-community==0.0.28
python-slugify==8.0.1